MeasureHelper

+++ utilities/src/main/java/bpc/dis/utilities/MeasureHelper/MeasureHelper.java
@@ -5,7 +5,7 @@ import android.util.TypedValue;
 
 public class MeasureHelper {
 
-    public static float pixelToDip(Context context, int pixel) {
+    public static float pixelToDip(Context context, float pixel) {
         return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, pixel, context.getResources().getDisplayMetrics());
     }
